Fix embedded JSON and use arrays for collections
Objects from Mongo were being serialized as JSON encoded strings and
then passed into the response which was JSON encoded itself. This now
properly passes the object so the JSON response is encoded as a whole.

Resources that represent a collection of objects, e.g. all students were
being returned as objects with the id as the entry map. Now, the
collections are simply being returned as an array of objects.